moving state to module
