+++ Models/App.js
@@ -1,6 +1,7 @@
 import HomePage from 'Pages/HomePage';
 import EditHikePage from 'Pages/EditHikePage';
 import MockBackend from 'Services/MockBackend';
+import State from 'Models/State';
 
 class Hike {
 	constructor(id, name, location, distance, rating, comments) {
@@ -13,26 +14,6 @@ class Hike {
 	}
 }
 
-class Current {
-	constructor() {
-		this.list = [];
-	}
-
-	get pages() {
-		return this.list;
-	}
-
-	pushPage(p) {
-		this.list.push(p);
-	}
-
-	popPage() {
-		this.list.pop();
-	}
-}
-
-export var current = new Current();
-
 export default class App {
 	constructor() {
 		this.mockBackend = new MockBackend();
@@ -40,17 +21,15 @@ export default class App {
 		this.mockBackend.getHikes().then(hikes => {
 			this.hikes = hikes;
 		});
-		this.state = current = new Current();
-		current.pushPage(new HomePage());
-		// current.pages.push(new HomePage());
+		this.state = new State();
+		this.state.pushPage(new HomePage());
 	}
 
 	goToHike(arg) {
-		current.pushPage(new EditHikePage(arg.data));
-		// current.pages.push(new EditHikePage(arg.data));
+		this.state.pushPage(new EditHikePage(arg.data));
 	}
 
 	get pages() {
-		return current.pages;
+		return this.state.pages;
 	}
 }
